+++ test/hirefire/test_resource.rb
@@ -2,9 +2,9 @@
 
 require "test_helper"
 
-class HireFire::ResourceTest < Minitest::Test
+class HireFireTest < Minitest::Test
   def test_configure_yields_configuration
-    config = HireFire::Resource.configure { |config| config }
-    assert_equal config, HireFire::Resource.configuration
+    config = HireFire.configure { |config| config }
+    assert_equal config, HireFire.configuration
   end
 end